// types/config.d.ts
export interface GlobConfig {
  // Site title
  title: string;
  // Service interface url
  apiUrl: string;
  // Upload url
  uploadUrl?: string;
  //  Service interface url prefix
  urlPrefix?: string;
  // Project abbreviation
  shortName: string;
  // App Id
  appId: string;
  // mch_huifu
  mch_huifu: string;
  // mch_origin
  mch_origin: string;
}